+++ src/main/java/xyz/acrylicstyle/bedwars/generators/EmeraldGenerator.java
@@ -26,7 +26,7 @@ public class EmeraldGenerator implements Generator {
     }
 
     @Override
-    public int getGenerateTime() {
+    public double getGenerateTime() {
         return time;
     }
